requirements.txt
gradio>=4.0.0
huggingface-hub
llama-cpp-python==0.2.90